+++ derive/src/context/num/mod.rs
@@ -23,6 +23,8 @@ impl Context {
         } = self;
         let generics = self.with_features(&[
             parse_quote!(AssumePrimeModulus),
+            parse_quote!(Equality),
+            parse_quote!(Order),
             parse_quote!(PartialAddition),
             parse_quote!(PartialSubtraction),
             parse_quote!(PartialMultiplication),