+++ tests/sqrt.rs
@@ -1,9 +1,7 @@
-use modtype::{use_modtype, DefaultModType};
-
 #[test]
 fn mod2() {
-    #[use_modtype]
-    type F = DefaultModType<2u32>;
+    #[modtype::use_modtype]
+    type F = modtype::F<2u32>;
 
     assert_eq!(F(0).sqrt(), Some(F(0)));
     assert_eq!(F(1).sqrt(), Some(F(1)));
@@ -11,8 +9,8 @@ fn mod2() {
 
 #[test]
 fn mod41() {
-    #[use_modtype]
-    type F = DefaultModType<41u32>;
+    #[modtype::use_modtype]
+    type F = modtype::F<41u32>;
 
     let mut num_quadratics = 0;
 
@@ -28,8 +26,8 @@ fn mod41() {
 
 #[test]
 fn mod1000000007() {
-    #[use_modtype]
-    type F = DefaultModType<1_000_000_007u64>;
+    #[modtype::use_modtype]
+    type F = modtype::F<1_000_000_007u64>;
 
     assert_eq!(F(42).sqrt(), Some(F(82_681_419)));
 }